+++ spring-cloud-starter-tencent-polaris-router/src/main/java/com/tencent/cloud/polaris/router/scg/PolarisReactiveLoadBalancerClientFilter.java
@@ -35,7 +35,7 @@ import com.tencent.cloud.common.util.JacksonUtils;
 import com.tencent.cloud.polaris.router.PolarisRouterServiceInstanceListSupplier;
 import com.tencent.cloud.polaris.router.RouterConstants;
 import com.tencent.cloud.polaris.router.RouterRuleLabelResolver;
-import com.tencent.cloud.polaris.router.spi.RouterLabelResolver;
+import com.tencent.cloud.polaris.router.spi.SpringWebRouterLabelResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Mono;
@@ -86,14 +86,14 @@ public class PolarisReactiveLoadBalancerClientFilter extends ReactiveLoadBalance
 	private final LoadBalancerProperties loadBalancerProperties;
 	private final MetadataLocalProperties metadataLocalProperties;
 	private final RouterRuleLabelResolver routerRuleLabelResolver;
-	private final List<RouterLabelResolver> routerLabelResolvers;
+	private final List<SpringWebRouterLabelResolver> routerLabelResolvers;
 
 	public PolarisReactiveLoadBalancerClientFilter(LoadBalancerClientFactory clientFactory,
 			GatewayLoadBalancerProperties gatewayLoadBalancerProperties,
 			LoadBalancerProperties loadBalancerProperties,
 			MetadataLocalProperties metadataLocalProperties,
 			RouterRuleLabelResolver routerRuleLabelResolver,
-			List<RouterLabelResolver> routerLabelResolvers) {
+			List<SpringWebRouterLabelResolver> routerLabelResolvers) {
 		super(clientFactory, gatewayLoadBalancerProperties, loadBalancerProperties);
 
 		this.clientFactory = clientFactory;
@@ -226,7 +226,10 @@ public class PolarisReactiveLoadBalancerClientFilter extends ReactiveLoadBalance
 		Map<String, String> labels = new HashMap<>(metadataLocalProperties.getContent());
 
 		// labels from rule expression
-		Map<String, String> ruleExpressionLabels = getExpressionLabels(exchange, peerServiceName);
+		Set<String> expressionLabelKeys = routerRuleLabelResolver.getExpressionLabelKeys(MetadataContext.LOCAL_NAMESPACE,
+				MetadataContext.LOCAL_SERVICE, peerServiceName);
+
+		Map<String, String> ruleExpressionLabels = getExpressionLabels(exchange, expressionLabelKeys);
 		if (!CollectionUtils.isEmpty(ruleExpressionLabels)) {
 			labels.putAll(ruleExpressionLabels);
 		}
@@ -235,7 +238,7 @@ public class PolarisReactiveLoadBalancerClientFilter extends ReactiveLoadBalance
 		if (!CollectionUtils.isEmpty(routerLabelResolvers)) {
 			routerLabelResolvers.forEach(resolver -> {
 				try {
-					Map<String, String> customResolvedLabels = resolver.resolve(exchange);
+					Map<String, String> customResolvedLabels = resolver.resolve(exchange, expressionLabelKeys);
 					if (!CollectionUtils.isEmpty(customResolvedLabels)) {
 						labels.putAll(customResolvedLabels);
 					}
@@ -254,10 +257,7 @@ public class PolarisReactiveLoadBalancerClientFilter extends ReactiveLoadBalance
 		return labels;
 	}
 
-	private Map<String, String> getExpressionLabels(ServerWebExchange exchange, String peerServiceName) {
-		Set<String> labelKeys = routerRuleLabelResolver.getExpressionLabelKeys(MetadataContext.LOCAL_NAMESPACE,
-				MetadataContext.LOCAL_SERVICE, peerServiceName);
-
+	private Map<String, String> getExpressionLabels(ServerWebExchange exchange, Set<String> labelKeys) {
 		if (CollectionUtils.isEmpty(labelKeys)) {
 			return Collections.emptyMap();
 		}