+++ spring-cloud-tencent-examples/polaris-router-example/router-caller-service/src/main/java/com/tencent/cloud/polaris/router/example/CustomSpringWebRouterLabelResolver.java
@@ -13,7 +13,6 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
- *
  */
 
 package com.tencent.cloud.polaris.router.example;
@@ -23,34 +22,31 @@ import java.util.Map;
 import java.util.Set;
 
 import com.google.gson.Gson;
-import com.tencent.cloud.polaris.router.spi.FeignRouterLabelResolver;
-import feign.RequestTemplate;
+import com.tencent.cloud.polaris.router.spi.SpringWebRouterLabelResolver;
 
+import org.springframework.http.HttpRequest;
 import org.springframework.stereotype.Component;
 
 /**
- *
- * Customize the business tag information obtained from the request
- *
- *@author lepdou 2022-05-12
+ * Custom router label resolver for spring web request.
+ * @author lepdou 2022-07-20
  */
 @Component
-public class CustomRouterLabelResolver implements FeignRouterLabelResolver {
+public class CustomSpringWebRouterLabelResolver implements SpringWebRouterLabelResolver {
 	private final Gson gson = new Gson();
-
-	@Override
-	public Map<String, String> resolve(RequestTemplate requestTemplate, Set<String> expressionLabelKeys) {
-		Map<String, String> labels = new HashMap<>();
-
-		User user = gson.fromJson(new String(requestTemplate.body()), User.class);
-
-		labels.put("user", user.getName());
-
-		return labels;
-	}
 
 	@Override
 	public int getOrder() {
 		return 0;
 	}
+
+	@Override
+	public Map<String, String> resolve(HttpRequest request, byte[] body, Set<String> expressionLabelKeys) {
+		Map<String, String> labels = new HashMap<>();
+		User user = gson.fromJson(new String(body), User.class);
+
+		labels.put("user", user.getName());
+		return labels;
+	}
+
 }