+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - [Feature: delete implement ServiceInstance](https://github.com/Tencent/spring-cloud-tencent/pull/482)
 - [Bugfix: update byte-buddy scope test to compile](https://github.com/Tencent/spring-cloud-tencent/pull/497)
 - [Fix the code analysis error.](https://github.com/Tencent/spring-cloud-tencent/pull/499)
+- [Optimize router label resolver spi](https://github.com/Tencent/spring-cloud-tencent/pull/503)